+++ components/search/SearchResultsControl.vue
